# api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenRHP API",
    "version": "0.1.0-dev",
    "description": "OpenReverseHomeProxy. Authenticated API shared by the UI, CLI and agents. Development status is exposed by capabilities; unavailable network features fail preflight."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/capabilities": {
      "get": {
        "operationId": "capabilities",
        "summary": "Detected platform capabilities and explicit limitations",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        }
      }
    },
...
    "/operations": {
      "get": {
        "operationId": "operations",
        "summary": "Read bounded persistent operation status",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Operation"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        }
      }
    },
    "/operations/{id}": {
      "get": {
        "operationId": "operation",
        "summary": "Resume observation after disconnect",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/transactions/{id}": {
      "get": {
        "operationId": "transaction",
        "summary": "Read durable network transaction state",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "retryable"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "Operation": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "state",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "state": {
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time"
          },
          "revision": {
            "type": "integer"
          },
          "error_code": {
            "type": "string"
          },
          "result": {
            "type": "object"
          }
        }
      },
...
    }
  }
}